+++ sampledb/frontend/location_types.py
@@ -73,6 +73,7 @@ class LocationTypeForm(FlaskForm):
     enable_responsible_users = BooleanField()
     enable_instruments = BooleanField()
     show_location_log = BooleanField()
+    enable_capacities = BooleanField()
 
 
 def show_location_type_form(type_id: typing.Optional[int]):
@@ -108,6 +109,7 @@ def show_location_type_form(type_id: typing.Optional[int]):
             location_type_form.enable_object_assignments.data = location_type.enable_object_assignments
             location_type_form.enable_responsible_users.data = location_type.enable_responsible_users
             location_type_form.enable_instruments.data = location_type.enable_instruments
+            location_type_form.enable_capacities.data = location_type.enable_capacities
             location_type_form.show_location_log.data = location_type.show_location_log
             # set translated texts from existing location type
             for text_name in translated_texts:
@@ -126,6 +128,7 @@ def show_location_type_form(type_id: typing.Optional[int]):
             location_type_form.enable_responsible_users.data = False
             location_type_form.show_location_log.data = False
             location_type_form.enable_instruments.data = True
+            location_type_form.enable_capacities.data = False
     else:
         translation_language_ids = flask.request.form.getlist('translation-languages')
         translation_language_ids = {
@@ -152,6 +155,7 @@ def show_location_type_form(type_id: typing.Optional[int]):
                 enable_object_assignments=location_type_form.enable_object_assignments.data,
                 enable_responsible_users=location_type_form.enable_responsible_users.data,
                 enable_instruments=location_type_form.enable_instruments.data,
+                enable_capacities=location_type_form.enable_capacities.data,
                 show_location_log=location_type_form.show_location_log.data,
                 **translated_texts
             ).id
@@ -164,6 +168,7 @@ def show_location_type_form(type_id: typing.Optional[int]):
                 enable_object_assignments=location_type_form.enable_object_assignments.data,
                 enable_responsible_users=location_type_form.enable_responsible_users.data,
                 enable_instruments=location_type_form.enable_instruments.data,
+                enable_capacities=location_type_form.enable_capacities.data,
                 show_location_log=location_type_form.show_location_log.data,
                 **translated_texts
             )