+++ sampledb/logic/federation/location_types.py
@@ -71,6 +71,7 @@ def import_location_type(
                 enable_object_assignments=location_type_data['enable_object_assignments'],
                 enable_responsible_users=location_type_data['enable_responsible_users'],
                 enable_instruments=False,
+                enable_capacities=False,
                 show_location_log=location_type_data['show_location_log'],
             )
             fed_logs.update_location_type(location_type.id, component.id)
@@ -85,6 +86,7 @@ def import_location_type(
             enable_object_assignments=location_type_data['enable_object_assignments'],
             enable_responsible_users=location_type_data['enable_responsible_users'],
             enable_instruments=False,
+            enable_capacities=False,
             show_location_log=location_type_data['show_location_log'],
             fed_id=location_type_data['fed_id'],
             component_id=component_id,
@@ -170,6 +172,7 @@ def _get_or_create_location_type_id(
             enable_object_assignments=False,
             enable_responsible_users=False,
             enable_instruments=False,
+            enable_capacities=False,
             show_location_log=False,
             fed_id=location_type_data['location_type_id'],
             component_id=component_id